requirements.txt
ai2thor==1.0.1
certifi==2019.3.9
chardet==3.0.4
Click==7.0
decorator==4.4.0
Flask==1.0.2
h5py==2.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
networkx==2.2
numpy==1.16.2
Pillow==5.4.1
progressbar2==3.39.3
protobuf==3.7.0
python-utils==2.3.0
PyYAML==5.1
requests==2.21.0
scipy==1.2.1
setproctitle==1.1.10
six==1.12.0
tabulate==0.8.3
tensorboardX==1.6
tqdm==4.31.1
urllib3==1.24.2
Werkzeug==0.15.3
torch==1.4.0
torchvision==0.5.0
scikit-image==0.19.3
matplotlib
